test season 11 with increase overlap

+++ Block_wise_GPS_correction.py
@@ -2790,30 +2790,30 @@ class Patch:
 
 		if increase_size:
 			if p1_x1>0+settings.PATCH_SIZE[1]/10:
-				p1_x1-=settings.PATCH_SIZE[1]/10
+				p1_x1-=int(settings.PATCH_SIZE[1]/10)
 
 			if p1_x2<9*settings.PATCH_SIZE[1]/10:
-				p1_x2+=settings.PATCH_SIZE[1]/10
+				p1_x2+=int(settings.PATCH_SIZE[1]/10)
 
 			if p1_y1>0+settings.PATCH_SIZE[0]/10:
-				p1_y1-=settings.PATCH_SIZE[0]/10
+				p1_y1-=int(settings.PATCH_SIZE[0]/10)
 
 			if p1_y2<9*settings.PATCH_SIZE[0]/10:
-				p1_y2+=settings.PATCH_SIZE[0]/10
+				p1_y2+=int(settings.PATCH_SIZE[0]/10)
 
 
 
 			if p2_x1>0+settings.PATCH_SIZE[1]/10:
-				p2_x1-=settings.PATCH_SIZE[1]/10
+				p2_x1-=int(settings.PATCH_SIZE[1]/10)
 
 			if p2_x2<9*settings.PATCH_SIZE[1]/10:
-				p2_x2+=settings.PATCH_SIZE[1]/10
+				p2_x2+=int(settings.PATCH_SIZE[1]/10)
 
 			if p2_y1>0+settings.PATCH_SIZE[0]/10:
-				p2_y1-=settings.PATCH_SIZE[0]/10
+				p2_y1-=int(settings.PATCH_SIZE[0]/10)
 
 			if p2_y2<9*settings.PATCH_SIZE[0]/10:
-				p2_y2+=settings.PATCH_SIZE[0]/10
+				p2_y2+=int(settings.PATCH_SIZE[0]/10)
 
 		return (p1_x1,p1_y1,p1_x2,p1_y2),(p2_x1,p2_y1,p2_x2,p2_y2)
